requirements.txt
asgiref==3.11.0
Django==5.1.4
sqlparse==0.5.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
djangorestframework>=3.16.0,<3.17
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.1
pycryptodome==3.19.0
Pillow==10.1.0
